seperating things out
